+++ app/Http/Controllers/LugarContactoController.php
@@ -39,7 +39,7 @@ class LugarContactoController extends Controller
      //funcion para insertar en la tabla
     public function store(Request $request)
     {
-        if(!$request->ajax())return redirect('/');
+        if(!$request->ajax() || Auth::user()->rol_id == 11)return redirect('/');
         $lugar_contacto = new Lugar_contacto();
         $lugar_contacto->nombre = $request->nombre;
         $lugar_contacto->usuario = Auth::user()->usuario;
@@ -49,7 +49,7 @@ class LugarContactoController extends Controller
     //funcion para actualizar los datos
     public function update(Request $request)
     {
-        if(!$request->ajax())return redirect('/');
+        if(!$request->ajax() || Auth::user()->rol_id == 11)return redirect('/');
         //FindOrFail se utiliza para buscar lo que recibe de argumento
         $lugar_contacto = Lugar_contacto::findOrFail($request->id);
         $lugar_contacto->nombre = $request->nombre;
@@ -58,7 +58,7 @@ class LugarContactoController extends Controller
 
     public function destroy(Request $request)
     {
-        if(!$request->ajax())return redirect('/');
+        if(!$request->ajax() || Auth::user()->rol_id == 11)return redirect('/');
         $lugar_contacto = Lugar_contacto::findOrFail($request->id);
         $lugar_contacto->delete();
     }
